+++ components/gallery/GalleryItemAction/GalleryItemActionCta.vue
