+++ components/gallery/GalleryItemAction/GalleryItemActionType/GalleryItemBuy.vue
@@ -26,8 +26,8 @@
 import { NeoButton } from '@kodadot1/brick'
 
 import { useGalleryItem } from '../../useGalleryItem'
-import GalleryItemPriceSection from '../GalleryItemPriceSection.vue'
-import GalleryItemPriceAction from '../GalleryItemPriceAction.vue'
+import GalleryItemPriceSection from '../GalleryItemActionSection.vue'
+import GalleryItemPriceAction from '../GalleryItemActionCta.vue'
 
 const { urlPrefix } = usePrefix()
 const { nft } = useGalleryItem()
